#125 fix clang compilation -Wold-style-cast

+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelTwoWindingsTransformer.cpp
@@ -1453,7 +1453,7 @@ ModelTwoWindingsTransformer::setGequations(std::map<int, std::string>& gEquation
   }
 
 
-  assert(gEquationIndex.size() == (unsigned int) sizeG_ && "Model TwoWindingsTransformer: gEquationIndex.size() != sizeG_");
+  assert(gEquationIndex.size() == static_cast<size_t>(sizeG_) && "Model TwoWindingsTransformer: gEquationIndex.size() != sizeG_");
 }
 
 double
